# app/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Construction Dashboard API",
    "description": "API for Construction Project Management Dashboard",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/projects/{id}/cost/change-orders": {
      "get": {
        "tags": [
          "Cost"
        ],
        "summary": "Get change orders for project",
        "description": "Returns all change orders for a specific project",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "type": "integer"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/ChangeOrderList"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
    "/projects/{id}/cost/invoices": {
      "get": {
        "tags": [
          "Cost"
        ],
        "summary": "Get invoices for project",
        "description": "Returns all invoices for a specific project",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "type": "integer"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/InvoiceList"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ChangeOrderList": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ChangeOrderSummary"
          }
        },
        "count": {
          "type": "integer"
        }
      }
    },
    "ChangeOrderSummary": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "number": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "float"
        },
        "status": {
          "type": "string"
        },
        "date_issued": {
          "type": "string",
          "format": "date"
        },
        "date_approved": {
          "type": "string",
          "format": "date"
        }
      }
    },
    "InvoiceList": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InvoiceSummary"
          }
        },
        "count": {
          "type": "integer"
        }
      }
    },
    "InvoiceSummary": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "invoice_number": {
          "type": "string"
        },
        "vendor": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "float"
        },
        "status": {
          "type": "string"
        },
        "invoice_date": {
          "type": "string",
          "format": "date"
        },
        "due_date": {
          "type": "string",
          "format": "date"
        }
      }
    },
...
  }
}